SARIF output format improvements #1739 #1740
