+++ .vscode/settings.json
@@ -95,7 +95,9 @@
     "deserializes",
     "Gitter",
     "hashtable",
+    "HEADREF",
     "Kubernetes",
+    "Multiformat",
     "Newtonsoft",
     "NOTCOUNT",
     "proxied",
@@ -103,12 +105,14 @@
     "PSRULE",
     "pwsh",
     "quickstart",
+    "REPOROOT",
     "runspace",
     "runspaces",
     "SARIF",
     "SBOM",
     "subselector",
     "unencrypted",
+    "Worktree",
     "xunit"
   ],
   "[csharp]": {
