+++ src/PSRule/Common/HashAlgorithmExtensions.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation.
+// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
 using System.Security.Cryptography;
@@ -11,5 +11,23 @@ internal static class HashAlgorithmExtensions
     {
         var hash = algorithm.ComputeHash(buffer);
         return string.Join("", hash.Select(b => b.ToString("x2")).ToArray());
+    }
+
+    public static string GetFileDigest(this HashAlgorithm algorithm, string path)
+    {
+        return algorithm.GetDigest(File.ReadAllBytes(path));
+    }
+
+    public static HashAlgorithm GetHashAlgorithm(this Options.HashAlgorithm algorithm)
+    {
+        if (algorithm == Options.HashAlgorithm.SHA256)
+            return SHA256.Create();
+
+        return algorithm == Options.HashAlgorithm.SHA384 ? SHA384.Create() : SHA512.Create();
+    }
+
+    public static string GetHashAlgorithmName(this Options.HashAlgorithm algorithm)
+    {
+        return algorithm == Options.HashAlgorithm.SHA256 ? "sha-256" : algorithm == Options.HashAlgorithm.SHA384 ? "sha-384" : "sha-512";
     }
 }
