+++ src/PSRule/Common/JsonConverters.cs
@@ -536,7 +536,7 @@ internal sealed class FieldMapJsonConverter : JsonConverter
 {
     public override bool CanRead => true;
 
-    public override bool CanWrite => false;
+    public override bool CanWrite => true;
 
     public override bool CanConvert(Type objectType)
     {
@@ -552,7 +552,15 @@ internal sealed class FieldMapJsonConverter : JsonConverter
 
     public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
     {
-        throw new NotImplementedException();
+        if (value is not FieldMap map) return;
+
+        writer.WriteStartObject();
+        foreach (var field in map)
+        {
+            writer.WritePropertyName(field.Key);
+            serializer.Serialize(writer, field.Value);
+        }
+        writer.WriteEndObject();
     }
 
     private static void ReadFieldMap(FieldMap map, JsonReader reader)
